requirements.txt
beautifulsoup4==4.12.3
Flask==3.1.0
numpy==2.1.3
python-dotenv==1.0.1
Requests==2.32.3
torch==2.5.1
transformers==4.46.2
yfinance==0.2.48
boto3==1.35.60